Implement user validation on farm updating

+++ CropCare/CropCare/Views/FarmSettingsPage.xaml.cs
@@ -57,6 +57,11 @@ public partial class FarmSettingsPage : ContentPage
             return;
         }
 
+        bool answer = await DisplayAlert("Confirm Update", "Are you sure you want to update this farm?", "Yes", "No");
+
+        if (!answer)
+            return;
+
         this.Farm.IconPath = Constants.Constants.GetIconPath(IconPicker.SelectedItem as string);
         this.Farm.Name = this.FarmName;
         await App.Repo.FarmsDb.UpdateItemAsync(this.Farm);
@@ -71,6 +76,12 @@ public partial class FarmSettingsPage : ContentPage
             await DisplayAlert("No Internet", "Please check your internet connection", "OK");
             return;
         }
+
+        bool answer = await DisplayAlert("Confirm Deletion", "Are you sure you want to delete this farm?", "Yes", "No");
+
+        if (!answer)
+            return;
+
         try
         {
             await App.Repo.UserToFarmDb.DeleteItemAsync(new UserToFarm(App.CurrentUser.Key, this.Farm.Key));
